feat(ingame_menu): only spawn newly clicked dropdowns
